hash_table REFACTOR simplify collision management
The current implementation manages collision by browsing the next
records to find an unused one.

This has the following consequences:
- when there is a lot of collisions, the insertion can take a lot
  of time before finding an empty entry.
- this may require to sometimes rehash the table to get rid of invalid
  records
- the code that handles the collisions is not trivial

This commit reworks the hash table to use a per-bucket list of records.
The insertion in a bucket can now be in O(1) if we don't have to check
for duplicates (see next commit).

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>