hash_table UPDATE insert element at the end of the bucket
----------
XXX Not sure why we would need this, but it appears that it solves
the validation test.
----------

Change the type of bucket list: instead of only referencing the first
record, reference both first and last records. Therefore we can add
new elements at the tail of the list.

This impacts how the records of a bucket will be browsed in case of
collisions:
- before this commit: last inserted is browsed first
- after this commit: first inserted is browsed first

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>