hash_table FEATURE new api to insert without lookup
When lyht_insert() is called, it first lookup for an existing object
that matches the previously specified val_equal() callback. In this
situation, the callback is invoked with mod=1.

In case there is a lot of collisions, this check can take some time.

Introduce a new API that bypasses this lookup operation.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>