hash_table FEATURE new api to insert without lookup
When lyht_insert() is called, it first searches for an existing object
that matches the previously specified val_equal() callback. In this
situation, the callback is invoked with mod=1.

In case there are a lot of collisions, this check can take some time.

Introduce a new API that bypasses this lookup operation, it will be
used in a next commit to optimize keyless list elements insertions.

This API can already be used internally in case of hash table resize,
where we don't need to check for duplicates.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>